Updated support for pnmp and yarn
