+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@github/local-action",
   "description": "Local Debugging for GitHub Actions",
-  "version": "3.0.5",
+  "version": "3.1.0",
   "type": "module",
   "author": "Nick Alteen <ncalteen@github.com>",
   "private": false,
